requirements.txt
Flask==3.0.2
pypdf==4.1.0
openai==1.16.2